Extract Store::record_failure() & helpers into impl modules.
Following in from 36292768b4257cb5aad82dba1925838922040101 this extracts
the nack impl further breaking down `Store`. Still a long way to go.
