+++ src/store/mod.rs
@@ -5,9 +5,11 @@ mod complete;
 mod complete_batcher;
 mod cron;
 mod enqueue_batcher;
+mod fail;
 mod group_committer;
 mod options;
 mod ready_index;
+mod requeue;
 mod results;
 mod scheduled_index;
 mod store;
